+++ test/test_miniruby.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'test_helper'
+require 'byebug'
 
 class TestMiniRuby < Minitest::Test
   def test_that_it_has_a_version_number