+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## [Unreleased]
 
+## [0.2.0] - 2026-04-15
+- Add new parser capabilities:
+  - Add `break`, `next`, `unless`
+  - modifier `if` and `unless`
+  - attribute access (`foo.bar`)
+
 ## [0.1.0] - 2025-02-12
 
 - Initial release